backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
scikit-learn>=1.5.0
numpy>=2.0.0
pandas>=2.2.2
joblib>=1.4.2
httpx>=0.27.0
reportlab>=4.0.0
